fix: move chip to display category and symmetric padding
- Move chip files from src/ui/general/chip/ to src/ui/display/chip/
  to match category reorg (chips are display elements, not general
  controls). Story title also updated to Components/Display/Chip.
- Symmetric horizontal padding: chip_content was 6/12/6/16 (asymmetric
  16L vs 12R). Plain text chips like "Basic" looked left-heavy.
  Changed to 6/12 symmetric. Leading/trailing icon overrides
  (has_leading/has_trailing) already reduce respective side to 8px.
